+++ my-app-test/src/components/UI/pages/About.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const About = () => {
+    return (
+        <h1>
+            Todo-List приложение созданное на React.
+        </h1>
+    );
+};
+
+export default About;
